Some cleaning up
