+++ FALdetector_files/global_classifier.py
