+++ FALdetector_files/local_detector.py
